% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.garma_model}
\alias{AIC.garma_model}
\title{AIC for model}
\usage{
\method{AIC}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
(double) Approximate AIC - uses approximation of whichever methoid is used to find model params.
}
\description{
AIC for model if available.
}
\details{
Note - like the R "arima" function, if a xreg parameter is supplied, the Log Likelihood and the AIC
do not include any component of the regression. It relates strictly to the residuals from
the regression which are used to fit the GARMA model.
}
