% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.garma_model}
\alias{autoplot.garma_model}
\title{ggplot of the Forecasts of the model.}
\usage{
\method{autoplot}{garma_model}(object, h = 24, include_fitted = FALSE, ...)
}
\arguments{
\item{object}{(garma_model) The garma_model from which to ggplot the values.}

\item{h}{(int) The number of time periods to predict ahead. Default: 24}

\item{include_fitted}{(bool) whether to include the 1-step ahead 'fitted' values in the plot. Default: FALSE}

\item{...}{other parameters passed to ggplot.}
}
\value{
A ggplot2 "ggplot" object. Note that the standard ggplot2 "+" notation can be used to enhance the default output.
}
\description{
The ggplot function generates a ggplot of actuals and predicted values for a "garma_model" object.
This adds in sensible titles etc as best it can determine.
}
\examples{
library(ggplot2)

data(AirPassengers)
ap <- as.numeric(diff(AirPassengers, 12))
mdl <- garma(ap, order = c(9, 1, 0), k = 0, method = "CSS", include.mean = FALSE)
autoplot(mdl)
}
