% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.garma_model}
\alias{coef.garma_model}
\title{Model Coefficients}
\usage{
\method{coef}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
(double) array of parameter value estimates from the fitted model.
}
\description{
Model Coefficients/parameters.
}
