% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval-Generic.R
\name{ConfidenceInterval}
\alias{ConfidenceInterval}
\title{ConfidenceInterval  class generic function.}
\usage{
ConfidenceInterval(garma.boot, ...)
}
\arguments{
\item{garma.boot}{An object of the \bold{GarmaSimBoot} 
class, as provided by \code{\link{GarmaSimBoot}}.}

\item{\dots}{Further arguments that specify the 
ConfidenceInterval object, check the \bold{slots}
arguments.}
}
\description{
A generic function working as a class constructor for 
the ConfidenceInterva class.
}
\details{
This generic function is designed to be a constructor for the
ConfidenceInterva class.
}
\author{
Matheus de Vasconcellos Barroso
}
