% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceInterval-Generic.R
\name{coverage}
\alias{coverage}
\title{coverage generic function.}
\usage{
coverage(object)
}
\arguments{
\item{object}{An object of the \bold{ConfidenceInterval} 
class, as provided by \code{\link{ConfidenceInterval}}.}
}
\description{
A generic function working to compute ci's coverage rates
}
\details{
This generic function is designed to extract the coverage
rates from a \code{\link{ConfidenceInterval}} object.
}
\author{
Matheus de Vasconcellos Barroso
}
