% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{css.ggbr.obj}
\alias{css.ggbr.obj}
\title{Estimate a Ggbr model using a CSS (Conditional Sum-of-Squares) method.
css.ggbr.obj - objective function to be minimised to get CSS estimates.
called from function "garma"}
\usage{
css.ggbr.obj(par, params)
}
\arguments{
\item{par}{- the parameters to evaluate the function at}

\item{params}{- other parameters - including the p, q, k, and scale parameters and (ss) the spectrum .}
}
\value{
The value of the objective at the point par.
}
\description{
Estimate a Ggbr model using a CSS (Conditional Sum-of-Squares) method.
css.ggbr.obj - objective function to be minimised to get CSS estimates.
called from function "garma"
}
