% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wll.R
\name{.wll.ggbr.obj}
\alias{.wll.ggbr.obj}
\title{Estimate a Ggbr model using the Whittle-like-Log (WLL) method.}
\usage{
.wll.ggbr.obj(par, params)
}
\arguments{
\item{par}{- the parameters to evaluate the function at}

\item{params}{- other parameters - including the p, q, k, and scale parameters and (ss) the spectrum .}
}
\value{
The value of the objective at the point par.
}
\description{
wll.ggbr.obj - objective function to be minimised to get WLL estimates.
It is not anticipated that an end-ser would have a need to call this function directly.
}
