% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{fitted.garma_forecast}
\alias{fitted.garma_forecast}
\title{Return the fitted values for a GARMA forecast.}
\usage{
\method{fitted}{garma_forecast}(object, ...)
}
\arguments{
\item{object}{(garma_model) A 'fit' object from the `garma()` function.}

\item{...}{other parameters, which are ignored.}
}
\description{
Return the fitted values for a GARMA forecast.
}
