% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaSpec.R
\docType{class}
\name{GammaSpec-class}
\alias{GammaSpec-class}
\title{\emph{Gamma}-GARMA specification class.}
\description{
An S4 class to represent a \emph{Gamma} garma specification object.
}
\details{
This is S4 class defines the basic structre of a \emph{Gamma} GARMA
speficication object.
}
\section{Slots}{

\describe{
\item{\code{sigma2}}{Numeric, specifying the \bold{sigma2} parameter.}

\item{\code{mu0}}{A numeric vector with length equal to the \bold{max.order}
of the \emph{Gamma}-GARMA model. See \code{\link{GarmaSpec}} for
further details.}
}}

\author{
Matheus de Vasconcellos Barroso
}
