\name{garma-package}
\alias{garma-package}
\alias{garma}
\docType{package}
\title{
\packageTitle{garma}
}
\description{
\packageDescription{garma}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{garma}
\packageIndices{garma}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{garma}

Maintainer: \packageMaintainer{garma}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
