% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaFit-Generic.R
\name{GarmaFit}
\alias{GarmaFit}
\title{GARMA fit  class generic function.}
\usage{
GarmaFit(garma, ...)
}
\arguments{
\item{garma}{An object of the \bold{GarmaSim} class, as provided
by \code{\link{GarmaSim}}.}

\item{\dots}{Further arguments that specify the GarmaFit object,
check the \bold{slots} arguments.}
}
\description{
A generic function working as a class constructor for the GarmaFit class.
}
\details{
This generic function is designed to be a constructor for the
GarmaFit class.
}
