% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaSim.R, R/GarmaSim-Generic.R
\docType{class}
\name{GarmaSim-class}
\alias{GarmaSim-class}
\alias{GarmaSim,ANY-method}
\alias{print,GarmaSim-method}
\alias{plot,GarmaSim-method}
\alias{summary,GarmaSim-method}
\title{\emph{GARMA} simulation class.}
\usage{
\S4method{GarmaSim}{ANY}(spec, ...)

\S4method{print}{GarmaSim}(x)

\S4method{plot}{GarmaSim}(x, confInt = 0.95, ...)

\S4method{summary}{GarmaSim}(object, ...)
}
\arguments{
\item{spec}{An object of the GarmaSpec class, as provided
by \code{\link{GarmaSpec}}.}

\item{\dots}{Further arguments that specify the simulation object.
Check the \bold{Slots} for further details.}

\item{x}{An object of the GarmaSim class, as provided
by \code{\link{GarmaSim}}.}

\item{confInt}{A numeric value, \eqn{0 < confInt <1},
giving the desired confidente interval size for the plot.}

\item{object}{An object of the GarmaSim class, as provided
by \code{\link{GarmaSim}}.}

\item{x}{An object of the GarmaSim class, as provided
by \code{\link{GarmaSim}}.}
}
\description{
An S4 class to represent a \emph{garma} simulation object.
}
\details{
This S4 class defines the basic structre of a \emph{GARMA}
simulation object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{GarmaSim}: GarmaSim Class constructor.

A method to create an object of the GarmaSim class.

This method creates an object of the GarmaSim class.

\item \code{print}: Print method for a GarmaSim object.

\item \code{plot}: Print method for a GarmaSim object default plot.

\item \code{summary}: Summary method for a GarmaSim object default plot.
}}

\section{Slots}{

\describe{
\item{\code{spec}}{An object of the \bold{GarmaSpec} class, as provided
by \code{\link{GarmaSpec}}.}

\item{\code{nmonte}}{A positive integer, specifying the number of Monte Carlo
simulations to perform, the default value is \bold{1000}.}

\item{\code{nsteps}}{A numeric vector with the number of steps in the
Garma model simulation, that is, the length of the time series
to simulate. The default value is \bold{100}.}

\item{\code{burnin}}{A numeric vector indicating the number of burn in
observations. If you want to generate only \code{nsteps} this
argument should be set equal to zero. Otherwise, provide a positive
integer. The default value is \bold{1000}.}

\item{\code{allow.parallel}}{Logical \bold{TRUE/FALSE} indicating
whether parallel computation via the foreach package
should be used. The default value is \code{TRUE}. OBS:paralllel
 backend must be registered prior to calling \code{\link{GarmaSim}}.}

\item{\code{seed}}{Numeric, the seed to \code{set.seed()} for
replicable examples. Default value is \bold{123}.}
}}

\author{
Matheus de Vasconcellos Barroso
}
