% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaSim-Generic.R
\name{GarmaSim}
\alias{GarmaSim}
\title{GARMA simulation class generic function.}
\usage{
GarmaSim(spec, ...)
}
\arguments{
\item{spec}{An object of the GarmaSpec class, as provided
by \code{\link{GarmaSpec}}.}

\item{\dots}{Further arguments that specify the GarmaSim object,
check the \bold{slots} arguments.}
}
\description{
A generic function working as a class constructor for the GarmaSim class.
}
\details{
This generic function is designed to be a constructor for the
GarmaSim class.
}
\author{
Matheus de Vasconcellos Barroso
}
