% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaSpec.R, R/GarmaSpec-Generic.R
\docType{class}
\name{GarmaSpec-class}
\alias{GarmaSpec-class}
\alias{GarmaSpec,ANY-method}
\title{GARMA specification class.}
\usage{
\S4method{GarmaSpec}{ANY}(family, ...)
}
\arguments{
\item{family}{A character vector specifying the
family of the specification object. Accepted values
are: "Po" for poisson and "GA"for gamma families.}

\item{\dots}{Further arguments that specify the model.}
}
\description{
An S4 class to represent a garma specification object.
}
\details{
This S4 class defines the basic structre of a GARMA
speficication object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{GarmaSpec}: GarmaSpec Class constructor.

A method to create an object of the GarmaSpec class.

This method creates an object of the GarmaSpec class.
}}

\section{Slots}{

\describe{
\item{\code{family}}{A character vector specifying the
family of the specification object. Accepted values
are: "Po" for poisson and "GA"for gamma families.}

\item{\code{beta.x}}{A numeric vector with length of
the desired specification.}

\item{\code{phi}}{A numeric vector with length of
the desired autoregressive term order specification.}

\item{\code{theta}}{A numeric vector with length of
the desired moving-average term order specification.}

\item{\code{X}}{A n x m matrix, where \code{n = nsteps + burnin + max.order}
and  \code{m = length(\strong{beta.x})}. Where \strong{nsteps}
is the number of simulations desired, \strong{burnin}
the number of burnin observations and \strong{max.order} is
equal to the highest order of the GARMA model (i.e.
\code{max(length(phi),length(theta)}.)}
}}

