% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaSpec-Generic.R
\name{GarmaSpec}
\alias{GarmaSpec}
\title{GARMA specification class generic function.}
\usage{
GarmaSpec(family, ...)
}
\arguments{
\item{family}{A character vector specifying the
family of the specification object. Accepted values
are: "Po" for poisson and "GA"for gamma families.}

\item{\dots}{Further arguments that specify the model,
check the \bold{slots} arguments.}
}
\description{
A generic function working as a class constructor for the GarmaSpec class.
}
\details{
This generic function is designed to be a constructor for the
GarmaSpec family.
}
\author{
Matheus de Vasconcellos Barroso
}
