% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipara.R
\name{ggbr_semipara}
\alias{ggbr_semipara}
\title{For a k=1 Gegenbauer process, use semi-parametric methods to estimate the Gegenbauer frequency and fractional differencing.}
\usage{
ggbr_semipara(x, alpha = 0.8, method = "gsp")
}
\arguments{
\item{x}{(num) This should be a numeric vector representing the process to estimate.}

\item{alpha}{(num)}

\item{method}{(char) One of "gsp" or "lpr" - lpr is the log-periodogram-regression technique, "gsp" is the Gaussian semi-parametric technique. "gsp" is the default. Refer Arteche (1998).}
}
\value{
An object of class "garma_semipara".
}
\description{
For a k=1 Gegenbauer process, use semi-parametric methods to estimate the Gegenbauer frequency and fractional differencing.
}
\examples{
data(AirPassengers)
ap <- as.numeric(diff(AirPassengers,12))
sp <- ggbr_semipara(ap)
print(sp)
}
