% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdiag.R
\name{gof}
\alias{gof}
\title{Goodness-of-Fit test for a garma_model.}
\usage{
gof(object)
}
\arguments{
\item{object}{(garma_model) The garma_model to produce the diagnostic plots for.}
}
\value{
None.
}
\description{
Provides a goodness-of-fit test for a GARMA Model, using the method of Fay & Philippe (2002).
}
\examples{
data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=FALSE)
gof(mdl)
}
