% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.garma_model}
\alias{logLik.garma_model}
\title{Log Likelihood}
\usage{
\method{logLik}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
Object of class "logLik" with values for the (approx) log-likelihood for the model
}
\description{
Log Likelihood, or approximate likelihood depending on the method.
}
\details{
Note - like the R "arima" function, if a xreg parameter is supplied, the Log Likelihood
does not include any component of the regression. It relates strictly to the residuals from
the regression which are used to fit the GARMA model.
}
