% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.garma_model}
\alias{plot.garma_model}
\title{The plot function generates a plot of actuals and predicted values for a "garma_model" object.}
\usage{
\method{plot}{garma_model}(mdl, h = 24, ...)
}
\arguments{
\item{mdl}{(garma_model) The garma_model from which to plot the values.}

\item{h}{(int) The number of time periods to predict ahead. Default: 24}

\item{...}{other arguments to be passed to the "plot" function.}
}
\value{
An R "plot" object.
}
\description{
The plot function generates a plot of actuals and predicted values for a "garma_model" object.
}
\examples{
data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=F)
plot(mdl)
}
