% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonSpec.R
\docType{class}
\name{PoissonSpec-class}
\alias{PoissonSpec-class}
\title{\emph{Poisson}-GARMA specification class.}
\description{
An S4 class to represent a \emph{Poisson} garma specification object.
}
\details{
This is S4 class defines the basic structre of a \emph{Poisson} GARMA
speficication object.
}
\section{Slots}{

\describe{
\item{\code{family}}{A character with the tag "PO".}

\item{\code{alpha}}{Numeric, specifying the \bold{offset} parameter.}

\item{\code{mu0}}{A numeric vector with length equal to the \bold{max.order}
of the \emph{Poisson}-GARMA model. See \code{\link{GarmaSpec}} for
further details.}
}}

\author{
Matheus de Vasconcellos Barroso
}
