% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garma_main.R
\name{predict2}
\alias{predict2}
\title{Predict2 future values.}
\usage{
predict2(object, n.ahead = 1)
}
\arguments{
\item{object}{(garma_model) The garma_model from which to predict the values.}

\item{n.ahead}{(int) The number of time periods to predict ahead. Default: 1}
}
\value{
A "ts" object containing the requested forecasts.
}
\description{
Predict ahead using algorithm of (2009) Godet, F
"Linear prediction of long-range dependent time series", ESAIM: PS 13 115-134.
DOI: 10.1051/ps:2008015
}
