% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.garma_model}
\alias{print.garma_model}
\title{print a garma_model object.}
\usage{
\method{print}{garma_model}(x, ...)
}
\arguments{
\item{x}{(garma_model) The garma_model from which to print the values.}

\item{...}{Other arguments. Ignored.}
}
\description{
The print function prints a summary of a "garma_model" object.
}
\examples{
data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=FALSE)
print(mdl)
}
