% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipara.R
\name{print.garma_semipara}
\alias{print.garma_semipara}
\title{Print a semiparameteric Gegenbauer estimation object.}
\usage{
\method{print}{garma_semipara}(x, ...)
}
\arguments{
\item{x}{An object of class garma_semipara.}

\item{...}{further parameters for print function}
}
\description{
Print a semiparameteric Gegenbauer estimation object.
}
