% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{qml.ggbr.obj}
\alias{qml.ggbr.obj}
\title{qml.ggbr.obj - objective function to be minimised to get QML estimates.
qml.ggbr.se2 - returns se estmates for parameters.
called from function "garma"}
\usage{
qml.ggbr.obj(par, params)
}
\arguments{
\item{par}{- the parameters to evaluate the function at}

\item{params}{- other parameters - including the p, q, k, and scale parameters and (ss) the spectrum .}
}
\value{
The value of the objective at the point par.
}
\description{
qml.ggbr.obj - objective function to be minimised to get QML estimates.
qml.ggbr.se2 - returns se estmates for parameters.
called from function "garma"
}
