% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{qml.ggbr.se2}
\alias{qml.ggbr.se2}
\title{qml.ggbr.se2 - calculate the se^2 figures for the parameter estimates}
\usage{
qml.ggbr.se2(par, params)
}
\arguments{
\item{par}{- the parameter estimates}

\item{params}{- other parameters - including the p, q, k, and scale parameters and (ss) the spectrum .}
}
\value{
The value of the objective at the point par.
}
\description{
qml.ggbr.se2 - calculate the se^2 figures for the parameter estimates
}
