% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qml.R
\name{qml_matricies}
\alias{qml_matricies}
\title{Estimate a Ggbr model using a QML (Quasi-Maximum-Likelihood) method.
qml_matricies - geterate the statespace matrices needed for QML method.}
\usage{
qml_matricies(par, p, q, k, include.mean)
}
\value{
- the QML matricies in a list.
}
\description{
Estimate a Ggbr model using a QML (Quasi-Maximum-Likelihood) method.
qml_matricies - geterate the statespace matrices needed for QML method.
}
