% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{residuals.garma_forecast}
\alias{residuals.garma_forecast}
\title{Return the residuals for a GARMA forecast.}
\usage{
\method{residuals}{garma_forecast}(object, ...)
}
\arguments{
\item{object}{(garma_model) A 'fit' object from the `garma()` function.}

\item{...}{other parameters, which are ignored.}
}
\description{
Return the residuals for a GARMA forecast.
}
