% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garma_main.R
\name{residuals.garma_model}
\alias{residuals.garma_model}
\title{Residuals}
\usage{
\method{residuals}{garma_model}(object, type = "response", h = 1, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{type}{(chr) The type of residuals. Must be 'response'.}

\item{h}{(int) The number of periods ahead for the residuals. Must be 1.}

\item{...}{Other parameters. Ignored.}
}
\description{
Response Residuals from the model.
}
