% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semipara.R
\name{semipara}
\alias{semipara}
\title{For a k=1 Gegenbauer process, use semi-parametric methods to estimate the Gegenbauer frequency and fractional differencing.
Then reverse to get a short memory process and return all this.}
\usage{
semipara(x)
}
\arguments{
\item{x}{(num) This should be a numeric vector representing the process to estimate.}
}
\value{
An object of class "garma_semipara".
}
\description{
For a k=1 Gegenbauer process, use semi-parametric methods to estimate the Gegenbauer frequency and fractional differencing.
Then reverse to get a short memory process and return all this.
}
\examples{
data(AirPassengers)
ap <- as.numeric(diff(AirPassengers,12))
sp <- semipara(ap)
print(sp)
}
