% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summary.garma_model}
\alias{summary.garma_model}
\title{The summary function provides a summary of a "garma_model" object.}
\usage{
\method{summary}{garma_model}(mdl, verbose = TRUE)
}
\arguments{
\item{mdl}{(garma_model) The garma_model from which to print the values.}

\item{verbose}{(bool) whether to print out the verbose version of the model or not. Default: TRUE}
}
\description{
The summary function provides a summary of a "garma_model" object.
}
\examples{
data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=F)
summary(mdl)
}
