% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.garma_model}
\alias{vcov.garma_model}
\title{Covariance matrix}
\usage{
\method{vcov}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
(double) estimated variance-covariance matrix of the parameter estimates
}
\description{
Covariance matrix of parameters if available
}
