% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whittle.R
\name{whittle.ggbr.obj}
\alias{whittle.ggbr.obj}
\title{Estimate a Ggbr model using the Whittle method.
whittle.ggbr.obj - objective function to be minimised to get Whittle estimates.
called from function "garma"}
\usage{
whittle.ggbr.obj(par, params)
}
\arguments{
\item{par}{- the parameters to evaluate the function at}

\item{params}{- other parameters - including the p, q, k, and scale parameters and (ss) the spectrum .}
}
\value{
The value of the objective at the point par.
}
\description{
Estimate a Ggbr model using the Whittle method.
whittle.ggbr.obj - objective function to be minimised to get Whittle estimates.
called from function "garma"
}
