% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garma_main.R
\name{fitted.garma_model}
\alias{fitted.garma_model}
\title{Fitted values}
\usage{
\method{fitted}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
(double) array of 1-step ahead fitted values for the model.
}
\description{
Fitted values are 1-step ahead predictions.
}
