% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaFit.R, R/GarmaFit-Generic.R,
%   R/GarmaSimBoot-Generic.R
\docType{class}
\name{GarmaFit-class}
\alias{GarmaFit-class}
\alias{GarmaFit,ANY-method}
\alias{print,GarmaFit-method}
\alias{plot,GarmaFit-method}
\alias{summary,GarmaFit-method}
\alias{plot,GarmaSimBoot-method}
\title{\emph{GARMA} simulation fit class.}
\usage{
\S4method{GarmaFit}{ANY}(garma, ...)

\S4method{print}{GarmaFit}(x)

\S4method{plot}{GarmaFit}(x, scales = "free", ...)

\S4method{summary}{GarmaFit}(object, ...)

\S4method{plot}{GarmaSimBoot}(x, scales = "free", ...)
}
\arguments{
\item{garma}{An object of the \bold{GarmaSim} class, as provided
by \code{\link{GarmaSim}}.}

\item{\dots}{Further arguments that specify the GarmaFit object,
check the \bold{slots} arguments.}

\item{x}{An object of the GarmaFit class, as provided
by \code{\link{GarmaFit}}.}

\item{scales}{A character specifying if the scales should be free or not.
The default value is \code{"free"} and accepted values are:
\code{'fixed', 'free_x', 'free_y' and 'free'}.}

\item{object}{An object of the GarmaFit class, as provided
by \code{\link{GarmaFit}}.}

\item{x}{An object of the GarmaFit class, as provided
by \code{\link{GarmaFit}}.}

\item{x}{An object of the GarmaFit class, as provided
by \code{\link{GarmaFit}}.}

\item{scales}{A character specifying if the scales should be free or not.
The default value is \code{"free"} and accepted values are:
\code{'fixed', 'free_x', 'free_y' and 'free'}.}
}
\description{
An S4 class to represent a \emph{garma} simulation fitted object.
}
\details{
This S4 class defines the basic structre of a \emph{GARMA}
simulation fitted object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{GarmaFit}: GarmaFit Class constructor.

A method to create an object of the GarmaFit class.

This method creates an object of the GarmaFit class.

\item \code{print}: Print method for a GarmaFit object.

\item \code{plot}: Plot method for a GarmaFit object.

\item \code{summary}: Summary method for a GarmaFit object.

\item \code{plot}: Plot method for a GarmaFit object.
}}

\section{Slots}{

\describe{
\item{\code{garma}}{An object of the \bold{GarmaSim} class, as provided
by \code{\link{GarmaSim}}.}

\item{\code{allow.parallel}}{Logical \bold{TRUE/FALSE} indicating
whether parallel computation via the foreach package
should be used. The default value is \code{TRUE}. OBS:paralllel
 backend must be registered prior to calling \code{\link{GarmaSim}}.}

\item{\code{seed}}{Numeric, the seed to \code{set.seed()} for
replicable examples. Default value is \bold{123}.}

\item{\code{errorhandling}}{Character, either 'try' or 'pass'}

\item{\code{n.try}}{Positive integer. If \code{errorhandling = 'try'},
this specifies the number of attempts in the algorithm.}

\item{\code{control}}{List. This is passed to the garmaFit2 function. The options
are given by  \code{\link[gamlss.util]{garmaFit}}.}

\item{\code{print.out}}{A matrix with values to be used by \code{print}.}

\item{\code{plot.out}}{A list with two elements: \describe{
  \item{db}{A data frame with XXXX}
  \item{db2}{A data frame with XXXX}
} with values to be used by \code{plot}.}

\item{\code{value}}{A list, where each element is a fitted
\emph{GARMA} object.}
}}

