% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GarmaSimBoot.R, R/GarmaSimBoot-Generic.R
\docType{class}
\name{GarmaSimBoot-class}
\alias{GarmaSimBoot-class}
\alias{GarmaSimBoot,ANY-method}
\alias{print,GarmaSimBoot-method}
\alias{summary,GarmaSimBoot-method}
\title{\emph{GARMA} simulation bootstrap class.}
\usage{
\S4method{GarmaSimBoot}{ANY}(sim, ...)

\S4method{print}{GarmaSimBoot}(x)

\S4method{summary}{GarmaSimBoot}(object, ...)
}
\arguments{
\item{sim}{An object of the \bold{GarmaSim} class, as provided
by \code{\link{GarmaSim}}.}

\item{\dots}{Further arguments that specify the GarmaSimBoot object,
check the \bold{slots} arguments.}

\item{x}{An object of the GarmaSimBoot class, as provided
by \code{\link{GarmaSimBoot}}.}

\item{object}{An object of the GarmaSimBoot class, as provided
by \code{\link{GarmaSimBoot}}.}
}
\description{
An S4 class to represent a \emph{garma} simulation bootstraped object.
}
\details{
This S4 class defines the basic structre of a \emph{GARMA}
simulation bootstraped object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{GarmaSimBoot}: GarmaSimBoot Class constructor.

A method to create an object of the GarmaSimBoot class.

This method creates an object of the GarmaSimBoot class.

\item \code{print}: Print method for a GarmaSimBoot object.

\item \code{summary}: Summary method for a GarmaSimBoot object.
}}

\section{Slots}{

\describe{
\item{\code{sim}}{An object of the \bold{GarmaSim} class, as provided
by \code{\link{GarmaSim}}.}

\item{\code{l}}{\code{l} is the fixed block length used in generating the replicate
time series}

\item{\code{R}}{A positive integer giving the number of
bootstrap replicates required.}

\item{\code{allow.parallel}}{Logical \bold{TRUE/FALSE} indicating
whether parallel computation via the foreach package
should be used. The default value is \code{TRUE}. OBS:paralllel
 backend must be registered prior to calling \code{\link{GarmaSimBoot}}.}

\item{\code{seed}}{Numeric, the seed to \code{set.seed()} for
replicable examples. Default value is \bold{123}.}

\item{\code{errorhandling}}{Character, either 'try' or 'pass'}

\item{\code{n.try}}{Positive integer. If \code{errorhandling = 'try'},
this specifies the number of attempts in the algorithm.}

\item{\code{boot.function}}{A function to summarise the bootstrap replicates.
The default function returns 0. Be aware that this is not a problem, as by
default the mean values is already being returned. This is useful if the user
wants to specify a quantity not being reported, as an example consider the
0.2 quantile.}

\item{\code{control}}{List. This is passed to the garmaFit2 function. The options
are given by  \code{\link[gamlss.util]{garmaFit}}.}

\item{\code{print.out}}{A matrix with values to be used by \code{print}.}

\item{\code{plot.out}}{A list with two elements: \describe{
  \item{db}{A data frame with XXXX}
  \item{db2}{A data frame with XXXX}
} with values to be used by \code{plot}.}

\item{\code{value}}{A list, where each element is a fitted
\emph{GARMA} object.}
}}

