% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof}
\alias{gof}
\title{Goodness-of-Fit test for a garma_model.}
\usage{
gof(object)
}
\arguments{
\item{object}{(garma_model) The garma_model to test.}
}
\value{
Invisibly returns the array of p-values from the test.
}
\description{
Provides a goodness-of-fit test for a GARMA Model, using Bartletts Tp test.
This has been justified for long memory and for GARMA models by Delgado, Hidalgo and Velasco (2005).
}
