% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garma_main.R
\name{logLik.garma_model}
\alias{logLik.garma_model}
\title{Log Likelihood}
\usage{
\method{logLik}{garma_model}(object, ...)
}
\arguments{
\item{object}{The garma_model object}

\item{...}{Other parameters. Ignored.}
}
\value{
Object of class "logLik" with values for the (approx) log-likelihood for the model
}
\description{
Log Likelihood, or approximate likelihood or part likelihood, depending on the method.
}
