% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.garma_model}
\alias{plot.garma_model}
\title{Plot Forecasts from model.}
\usage{
\method{plot}{garma_model}(x, ...)
}
\arguments{
\item{x}{(garma_model) The garma_model from which to plot the values.}

\item{...}{other arguments to be passed to the "plot" function, including h (int) - the number of periods ahead to forecast.}
}
\value{
An R "plot" object.
}
\description{
The plot function generates a plot of actuals and predicted values for a "garma_model" object.
}
\examples{
data(AirPassengers)
ap  <- as.numeric(diff(AirPassengers,12))
mdl <- garma(ap,order=c(9,1,0),k=0,method='CSS',include.mean=FALSE)
plot(mdl)
}
