% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbr_semipara.R
\name{print.ggbr_factors}
\alias{print.ggbr_factors}
\title{Print a 'ggbr_factors' object.}
\usage{
\method{print}{ggbr_factors}(x, ...)
}
\arguments{
\item{x}{An object of class ggbr_factors}

\item{...}{further parameters for print function}
}
\value{
null
}
\description{
Print a 'ggbr_factors' object.
}
